[Feature] Refactor ConvertTo-Json to expose JsonObject.ConvertToJson as public API
